+++ appmap/__init__.py
@@ -1,7 +1,7 @@
 """AppMap recorder for Python"""
 
 from ._implementation.env import Env # noqa; F401
-from ._implementation.recording import Recording  # noqa: F401
+from ._implementation.recording import Recording, instrument_module  # noqa: F401
 from ._implementation import generation  # noqa: F401